Improve Unicode keyboard event support
This adds support for some of the non-ASCII characters sent via
Unicode input events, as well as resolves an issue with ASCII
characters that require Shift to be typed in.

The main issue with handling RDP input events is that X11 allows
only to simulate physical key presses by injecting the scan-codes
(called KeyCodes in X11). Thus, in order to type in arbitrary
Unicode characters we need to determine which KeyCodes to inject.

The first step is to convert Unicode code points to KeySyms which
are the name of visible glyphs and metakeys in X11. We use a
hardcoded table mapping to lookup a KeySym associated with given
UCS values.

After that we can use XKeysymToKeycode() function to map a KeySym
to corresponding KeyCode. It uses X server internal mapping that
is used to support keyboard layouts. However, we cannot call this
function directly from the input event handler as the handler is
executed by the X server thread. Calling this function will result
in deadlock. Thus we spawn a helper thread to handle Unicode input
asynchronously.

Also, each keycode can have multiple keysyms associated with it.
The exact glyph typed in depends on the state of modifier keys.
Modern systems usually support only two keysyms: the first is used
when the Shift key is not pressed, and the second one is used when
the Shift key has been pressed. We handle this case and inject
additional Shift key presses and releases to type in the character
we want.

This solution is not perfect though and has a couple of caveats:

- Not all Unicode range is supported.

  Only glyphs that have keysyms are supported by this approach.
  For example, CJK ideograms do not have associated keysyms so
  it is not possible to input Chinese or Japanese text.
  Unfortunately, the clients must run an IME on the remote
  system to be able to use such languages.

- The remote desktop must use a suitable keyboard layout.

  For example, Russian keyboard layout must be activated in the
  remote session in order to input Russian characters via Unicode
  input events. Activating the keyboard layout modifies the server
  keycode<->keysym mapping that we use to convert Unicode input
  to keycodes, thus a proper mapping is required so that client
  applications correctly interpret the injected keycode events
  they receive from the X server.